pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "builderbench"
version = "1.0.0"
description = "BuilderBench: The Building Blocks of Intelligent Agents"
requires-python = ">=3.10"
dependencies = [
    "mujoco >= 3.1.6",
    "dm_control >= 1.0.20",
    "gymnasium[mujoco]",
    "openai",
    "anthropic",
    "google-genai",
    "simple-parsing",
    "mediapy",
    "tabulate",
    "tqdm",
    "scipy",
    "pynput",
]

[tool.setuptools.packages.find]
include = ["builderbench*", "agents*"]
exclude = ["rl*"]
